# speed up javac and error-prone tasks by using less resources