Speed up javac and error-prone tasks, by using less resources

closes #11924
closes #11925