@rmuir 2022-11-13
I will try it out. I tried something similar to this... maybe exactly this... but when i inspected the forked processes for error-prone with `ps` they weren't actually getting the flags.