@dweiss 2022-11-13
I can see jvm flags being dumped if I add ```-XX:+PrintFlagsFinal``` to this patch (and either force the fork mode or run with an alt. jvm, which has the same side-effect).