@rmuir 2022-11-14
i tried looking at the kotlin code but didnt see any obvious way to sneak args in there. Also not really fluent in kotlin lol