@dweiss 2022-11-14
I didn't look at kotlin code, since I'm not fluent there at all... :) But I had a hunch to just upgrade the plugin thinking it's probably a bug there... and it seems it has been fixed. I added -XX:+PrintFlagsFinal after upgrading and it seems to pick it up now:
```
gradlew -p lucene\core compileJava -Dvalidation.errorprone=true --rerun-tasks --info
```

results in:

![image](https://user-images.githubusercontent.com/199470/201738094-b0f656ce-a09f-4b82-b32a-9343106aa25a.png)