@rmuir 2022-11-14
definitely seems to be working as it is spamming my console :) Thank you!

I'll test do some simple benchmarks vs main with the `task.times`, error-prone on and off, and see where we're at. The fact I am typing right now in my browser while it is running is promising, previously this was not possible as it would overwhelm my machine :)