@dweiss 2022-11-14
Please feel free to hack anything on my branch or cherry pick what you need - this was just a poc experiment.