@rmuir 2022-11-14
Thanks, this definitely got me past my stumbling block with error-prone. I will do more tests and try to get it ready.

as a first benchmark i tried './gradlew clean && ./gradlew check -x test -Dvalidation.errorprone=true -Ptask.times=true`, it makes a huge difference...

with error-prone enabled:

main:
```
BUILD SUCCESSFUL in 8m 30s
736 actionable tasks: 732 executed, 4 up-to-date

Aggregate task times (possibly running in parallel!):
 550.91 sec.  compileJava
 384.45 sec.  compileTestJava
```

Patch:
```
BUILD SUCCESSFUL in 6m 50s
736 actionable tasks: 732 executed, 4 up-to-date

Aggregate task times (possibly running in parallel!):
 295.94 sec.  compileJava
 286.68 sec.  compileTestJava
```