@rmuir 2022-11-14
OK i sped up the non-errorprone case too. It is best not to fork at all, forking is not wanted or helpful (but when it happens, we want reasonable options set). Instead, tuning options should be applied to `org.gradle.jvmargs`. Personally I've been doing this for over a year already, but i wiped my customizations in order to benchmark and make it faster for everyone:

error-prone disabled:
main:
```
BUILD SUCCESSFUL in 5m 18s

Aggregate task times (possibly running in parallel!):
 100.78 sec.  compileJava
  73.49 sec.  compileTestJava
```

patch:
```
BUILD SUCCESSFUL in 4m 14s

Aggregate task times (possibly running in parallel!):
  65.63 sec.  compileJava
  60.94 sec.  compileTestJava
```