@rmuir 2022-11-14
note, if you want to try this branch out yourself, you need to be sure to `clear/regenerate` gradle.properties after switching branches to see the impact.