@rmuir 2022-11-14
I'd like to reduce the 3GB heap of gradle we set by default too. But let's just make that another issue/discussion as maybe it is more controversial. I've been overriding the 3GB gradle heap myself and setting `org.gradle.jvmargs=-Xms1g -Xmx1g ...` forever. I think the 3GB heap was a relic from when solr was also in the build?