@rmuir 2022-11-15
@dweiss a difference may be the hyperthreading. on my tiny 2-core, using the hyperthreads results in a real speed benefit over using the default `ncpu/2`
```
org.gradle.workers.max=4
tests.jvms=4
```

For users with the default `ncpu/2` there is extra headroom available. so if we wastefully spin up too many resources, it doesnt hurt them. I want to use all my hardware.