@rmuir 2022-11-15
@dweiss i'll run simple bench on some mac laptops i have before pushing. i know they are very commonly used.