@rmuir 2022-11-15
I tested on M1 mac that has 10 cores (8 performance + 2 efficiency). With these higher concurrency levels, the benchmark is a little flawed since there are a couple of single-task bottlenecks that dominate. Plus it is just so much faster than my computer. Still, I think it shows that there's less contention and less overwhelming the system? 

`org.gradle.workers.max=5` (the defaults generated for you by lucene build)
```
main: 1m 31s
patch: 1m 34s
```

`org.gradle.workers.max=8` (matching number of performance cores)
```
main: 1m 44s
patch: 1m 30s
```

`org.gradle.workers.max=10` (matching number of total cores)
```
main: 1m 47s
patch: 1m 32s
```